pyproject.toml
[project]
name = "uv-pl-template"
version = "0.1.0"
description = ""
license = "MIT"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
   "hydra-colorlog>=1.2.0",
   "hydra-core>=1.3.2",
   "lightning==2.5.0.post0",
   "rich>=13.9.4",
   "rootutils>=1.0.7",
   "torchvision>=0.20.1",
   "wandb>=0.19.1",
]

[dependency-groups]
dev = [
   "mypy>=1.14.0",
   "ruff>=0.8.4",
   "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
clean-logs = "rm -rf logs/**"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
ignore = ["B905", "D417", "E501", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
